Add banner for old versions
Sometimes you'll come across a link on Google or StackOverflow that points to
an old version of the docs. It's not obvious from the current site design when
you're on an old version of the docs.

This adds a banner to all pages other than the highest semver'd folder of docs.

Closes #20