Build
This patch should fix all the broken links in old version warning banner.

Not all the broken links are fixed by #45 and #43. #44 fixed the root cause of the bug, but the doc is not rebuilt after the it was merged.

This patch contains the compiled document from 7e8c8e3, and all the broken links should be fixed now.