chore: Switch to a maintained Markdown parser
namp is [archived](https://github.com/gjtorikian/namp). So switching to a more maintained and feature-rich Markdown parser. This is a preparation for implementing #47.

Had to put metadata between `--- ---` (Jekyll syntax) as the previous syntax was not supported by the new Markdown parser. Separate PR to land the same changes in the Karma repository is coming.

Besides the above, the output is pretty much the same. Only minor escaping/whitespace changes in the rendered website + a couple of pieces of text, which were missed because of the bug in the old markdown parser.