fix: remove `/karma/` from index and gitignore it
`/karma/` got tracked in 24a3ba6. The old .gitignore with the pattern `/karma/*` does not ignore the directory itself, only its contents. With the pattern `/karma/` we ignore the directory and its contents. We cannot have such directory when newly cloning the repository because the command at /sync-docs.sh:30 fails with "fatal: git checkout: --detach does not take a path argument".